backend/requirements.txt
# ─── Web framework ────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0          # Required for FastAPI WebSocket endpoints (Day 3)

# ─── Async Redis (redis.asyncio is built into redis v4+; do NOT add aioredis) ─
redis[hiredis]>=5.0.0

# ─── Kafka / Stream processing ────────────────────────────────────────────────
faust-streaming>=0.10.0
kafka-python>=2.0.0

# ─── ML / Data ────────────────────────────────────────────────────────────────
xgboost>=2.0.0
scikit-learn>=1.4.0
shap>=0.45.0
pandas>=2.0.0
numpy>=1.26.0
joblib>=1.3.0
networkx>=3.0

# ─── Config / Utils ───────────────────────────────────────────────────────────
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
